requirements.txt
matplotlib==2.2.2
numpy==1.14.5
opencv-python==3.4.0.14
Pillow==5.3.0
scikit-image==0.14.1
scipy==1.1.0
svgwrite==1.2.1

# local package
-e .

# came with the cookiecutter template
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
